Address review comments

+++ llmnr.h
@@ -22,7 +22,7 @@
 #include <stdbool.h>
 
 void llmnr_set_hostname(const char *hostname);
-void llmnr_init(const char *hostnames[], int hostname_count, bool ipv6);
+void llmnr_init(const char *hostnames[], size_t hostname_count, bool ipv6);
 void llmnr_release(void);
 void llmnr_recv(int sock);
